requirements.txt
astropy==5.1
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.5.0
scipy==1.9.3
setuptools==65.6.3
streamlit==1.19.0
